DICE: parse and stringify
